Start connecting only after a button is pressed.
Currently, the connection phase is a synchronous call, which ends
up blocking the UI for some time if the connection can't be established.
Before this commit, this was done in app startup, which resulted into
very bad user experience.

Moving the logic behind a button makes it also easier to handle connection
errors and let's the user retry connecting.